Add myself to the list of trusted users
I have read the entire README and given that I lately had to debug some
aarch64 stuff, it would have been great to have had access to such a
machine.

Even though it's a bit inconsistent, I added the path to my key using a
path without the ./ in front of it. The reason why I didn't do it for
all the other keys as well is to avoid conflicts.

Signed-off-by: aszlig <aszlig@nix.build>